requirements.txt
python-telegram-bot[webhooks]==22.3
python-dotenv
pillow
# openai
google-generativeai
assemblyai
requests
asyncpg